+++ packages/react/src/index.ts
@@ -2,6 +2,7 @@ export * from './ThemeProvider'
 export { useTheme, createUseStyles } from 'react-jss'
 
 // Components
+export * from './components/AppBarTop'
 export * from './components/Avatar'
 export * from './components/Badge'
 export * from './components/Button'